+++ common/src/main/java/net/mehvahdjukaar/every_compat/misc/AllWoodItem.java
@@ -3,6 +3,7 @@ package net.mehvahdjukaar.every_compat.misc;
 
 import net.mehvahdjukaar.moonlight.api.client.ICustomItemRendererProvider;
 import net.mehvahdjukaar.moonlight.api.client.ItemStackRenderer;
+import net.minecraft.util.datafix.DataFixers;
 import net.minecraft.world.item.Item;
 
 import java.util.function.Supplier;